requirements.txt
django==5.1.7
psycopg2-binary==2.9.10
python-decouple==3.8

djangorestframework==3.15.2
drf-spectacular==0.28.0

djangorestframework-simplejwt==5.5.0